requirements.txt
yt-dlp>=2023.12.30
pyinstaller>=6.0.0
Pillow>=10.0.0
